requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
httpx==0.27.2
pika==1.3.2
sqlalchemy==2.0.25
python-dotenv==1.0.1
requests==2.32.3
python-consul==1.1.0